remove strategy and add nat
