+++ src/vpn.h
@@ -23,22 +23,8 @@
 
 #include <time.h>
 
-#ifdef TARGET_WIN32
-#include "win32.h"
-#else
-#include <netdb.h>
-#include <netinet/in.h>
-#include <arpa/inet.h>
-#include <sys/socket.h>
-#endif
 #include "args.h"
-
-/* the structure to store known client addresses for the server */
-typedef struct {
-  struct sockaddr_storage addr;
-  socklen_t addrlen;
-  time_t last_recv_time;
-} addr_info_t;
+#include "nat.h"
 
 typedef struct {
   int running;
@@ -56,10 +42,6 @@ typedef struct {
 #endif
   unsigned char *tun_buf;
   unsigned char *udp_buf;
-
-  /* known client addrs for the server */
-  int nknown_addr;
-  addr_info_t *known_addrs;
 
   /* the address we currently use */
   struct sockaddr_storage remote_addr;
